made the navbar modular
